fix: fix TexTalkNode stack overflow
This was caused by a TexTalkNode having a `parent` attribute
with a attribute that refers back to the Node as a child.

This causes a stack overflow in the generted `toString()`,
`hashCode()` etc.  The solution was to not have the `parent`
as a parameter in a data class's primary constructor.