fix: fix two replacement bugs (#41)
When inline commands were processed, commands in an 'is' node
were processed, which is incorrect.  Now they are not.

The `Phase2.toCode()` methods have been updated to handle the
cases where the node doesn't have any ClauseList arguments.

Now replacing a `Represents:` group works as expected.  Previously,
it threw a `ClassCastException`.