feat: support textalk single sub/sup params
Now if a sub or a sup part of a command consists of only a
single identifier, curly braces do not need to be placed around
it.

That is, previously `a_{b}` would parse, but `a_b` would not.
Now, they are both supported.

In addition the `toCode` method does not print curly braces around
a sub or sup part of a command if and only if it contains a single
identifier.

That is, if `a_{b}` is parsed, and the `toCode` method is called
on the result, the output will be `a_b`.