chore: use flattened signatures (#96)
Now a command such as `\f[x, y]{a, b}` has a signature
`\f[]{}` instead of `\f[?,?]{?,?}`.  In addition, the command
`\f{x, y}{a, b}{c}...` has a signature `\f{}...`.  This is
done to easily support variadic arguments and groups.

In particular, commands are distinguished by their names and
shape (number of [] and {} groups and whether they contain a
variadic group {}...) instead of by all of those features as
well as the number of arguments in each group.