feat: now the chalktalk ast is immutable
Previously, the ast was not completely immutable because each
node had `row` and `column` properties that were mutable.  This
made it difficult to compare nodes with the same structure but
different row and/or column.

Now row and column information is stored in a separate
`LocationTracker` object and a new `MathLingua.parseWithLocations()`
method is available to get a `Document` and a `LocationTracker`.