feat: CLI supports outputing expanded MathLingua (#142)
The CLI has a new --output option that, when specified, outputs
the contents of the supplied files.  The option can have values
'mathlingua', 'html', or 'none' (default).

In addition, the --expand flag has been added.  If that flag is
provided with the --output option, the statements in the output
will have their form fully expanded into their equivalent LaTeX
representations.

Also note that when the html output option is provided, the
generated html code uses KaTeX to render the math in the
document but references the KaTeX from its cdn.  Thus the html
is fully self contained.