feat: formalize the use of () vs {}
A command signature of a Defines can contain at most one paren
group, and if provided, the Defines must define a function like
object whose parameters exactly match those of the signature.

For example, the following is valid:
```
[\f(x, y)]
Defines: f(x, y)
means: 'something'
```
while the following is invalid since the signature has one
parameter `x` while the target of the Defines has two parameters
`x` and `y`:
```
[\f(x)]
Defines: f(x, y)
means: 'something'
```

Note that the above two examples are when a Defines contains a
paren group.  If it doesn't, then the target of the Defines can
be a function of any form.  Thus the following are both valid:
```
[\function]
Defines: f(x)
means: 'something'

[\another.function]
Defines: f(x, y)
means: 'something'
```

With respect to matching, parens can be omitted and a match will
still occur.  That is, given the Defines:
```
[\Sin(x)]
Defines: f(x)
means: '...'
Metadata:
. written: "\sin(x?)"
```

Then `\Sin(y)` would be expanded as `\sin(y)` while `\Sin` would be
expanded as `\sin(x?)` to signify that the `x` parameter was not
specified.

This `\sin(x?)` form specifies that the expression should be
interpreted as a function `\sin` of one variable as compared to
`\sin(x)` which is the value of the `\sin` function at `x`.