feat: support multiple sections with the same name (#197)
In particular, the sections with the same name do not need
to be continuous.  That is, the following is supported:
```
Theorem:
given:
where:
then:
using:
where:
```
Notice that there are two `where:` sections.  The first is
accessed by `where` and the second by `where1`.

The use of `where` instead of `where0` for the first `where:`
section is used since most of the time sections are not duplicated.