feat!: results use the `Result:given:then:` structure
Specifically, a `Theorem:` is now of the form:
```
Theorem: "some name"
given: X, Y
where: 'X, Y is \something'
then:
. 'X + Y is \something'
using: +
where: '+ := \some.plus'
```
`Axiom:` and `Conjecture:` forms are similar.

This makes a theorem more clear as to what it assumes and provides
and is intended to make it easier to translate to other languages
like Lean.