feat: support `Foundation:` groups
Foundation groups are used to specify definitions that are builtin
based on the underlying system used to check MathLingua definitions
for consistency.

For example, if this underlying system is Lean, there could be a
Foundation group
```
[\function:on{A}to{B}]
Foundation: f(x)
when: 'A, B is \Type'
means:
. lean: "Specifies a function f? : A? \to B?"
```
that is used to explicitly specify that `\function:on{A}to{B}` is
treated specially and is treated as `f : A \to B` in the underlying
Lean code.

Note: Foundation groups are used to communicate to readers of
MathLingua code what signatures are builtin to the underlying
system.