chore!: all named groups must start with a colon
Prior to this change named groups would be written as:
```
\function:on{A}to{B}
```
Now they are written as:
```
\function:on{A}:to{B}
```
Note the second `:` is now required.

This change was done so that keywords can be used as named group
names.  That is, now `\set.complement:of{A}:in{U}` is allowed since
without this change, the `in` would be interpreted as a keyword.