feat: add `site:name:` group
The `site:name:` group can be used as a reference to specify a
website.  The HtmlCodeWriter creates an `<a>` element for a
url and truncates the length if it is too long.

For example,
```
Theorem:
...
Metadata:
. reference:
  . site: "https://en.wikipedia.org/wiki/Magma_(algebra)"
    name: "Magma"
```
will have the anchor's displayed value as
```
en.wikipedia.org/.../Magma_(algebra)
```

In addition, this change include improvements to the rendered
HTML to center the search box and to not zoom in on mobile
devices when entering text.