fix: `mlg render --stdout <filename>` works (#339)
There was a regression that caused `--stdout` to not output the
result to standard output.  In addition, passing a file, instead
of a directory, to `mlg render` would not correctly render the
file.  This caused the MathLingua Langua Support VS Code
extension to stop working.  Both of these issues have been
addressed.