feat!: simplify the `mlg render` cli options (#346)
Now `mlg render` excepts an optional file argument
(instead of a list of file arguments).

If no file is provided, all of the files in the
`content` directory are rendered into a single
dynamic HTML document.

Otherwise, if a file is processed, only that file's
contents are rendered.

In either case, all of the files in the `content`
directory are used as context for definitions to
use when rendering the file or files.