feat: check function signature arg count (#350)
Now if the signature of the `Defines:` group specifies a function
(specifically if the signature contains a parens) then a check
is done to make sure the defines specifies a function.

That is, if the signature is `\something(x, y)` then the
defines must define a function `f(x, y)` that has exactly the
two arguments `x` and `y`.  That is, the count and name of the
arguments must match.

Note that if the signature doesn't have parens, then the
defines can still define a function.  For example, a defines
`\bounded.function` that specifies a function `f(x)` with certain
properties is still valid.