chore: introduce the `VirtualFileSystem` (#355)
The `VirtualFileSystem` represents a filesystem that could either
be an in-memory system or the regular disk based system.  In
addition, a new `Logger` interface represents something that can
create logs and abstracts away the use of `println`.

With these, a new `Mathlingua` object exists that can do the
operations of the `mlg` cli (check, render, clean, etc.) using
a `VirtualFileSystem` and a `Logger` without being tightly
coupled to using the underlying real filesystem or standard out.

As a result, it is now easy to have an entire MathLingua collection
of content completely in memory or on disk and perform the actions
that can be performed by the cli.

This allows for a better API to interact with a MathLingua
collection since it does not require physically writing any files
to disk.