feat: support `toc` files (#380)
Now a `toc` file can exist in any directory or subdirectory of
the `content` directory.  Its contents list the names of files
or directories in the directory which it resides.

It a `toc` file exists in a directory, `mlg render` will display
items in the file list in the order in the `toc` file and will
include an item if and only if it is in the `toc` file.

Thus, `toc` files allow one to both ignore certain files as well
as ensure a particular ordering of files in the file list in the
rendered html.

If a directory does not contain a `toc` file, all files and
directories in that directory will be listed in alphabetical
order.