feat: support underscores in names and operators (#391)
Now identifiers and operators can contain underscores.  For
example, `a_0` or `*_1`.

Note that `a_{i}` is a function invocation using a subparam while
`a_i` is the identifier with name `a_i`.  This is similar to how
`f(i)` is a function call while `fi` is an identifier name.

Specifically, the `{}` in a subparam specifies a function call in
the same way that `()` indicates a function call.