feat: add `mlg serve`
Now `mlg serve [--port]` can be run from a MathLingua code
root to start a server listening on the specified port (that
defaults to 8080).  Whenever the server is visited, the code
is rerendered and the resulting html is returned.

In addition, any output (either of success or failure with errors)
is logged to the console.