Revert "chore: use multiplatform libs for I/O (#422)"
This reverts commit 1c926fd90784b4de373ca081c2b2ac41db750084.

A lot of additional work was needed to get Kotlin native to work
(experimented in the separate experiment-with-kotlin-native
branch).

Even then, all of the tests had to be deleted in the
experiment-with-kotlin-native branch since they would need
to use a completely different test library.  Also, some of the I/O
code was hand-written to work with my specific test code.

Parsing and rendering my sample MathLingua code took ~1.3 seconds
in Kotlin native and ~2.7 in Kotlin jvm.  However, build speed
in Kotlin native is too slow in Kotlin natvie compared to Kotlin
jvm.  Furthermore, the Intellij dev experience is not very good.
In particular, autocomplete doesn't work.

Thus, I am reverting this change to instead use Kotlin JVM.