fix: SourceCollection.removeSource() works correctly
There was a change that had the input to removeSource() be the
relative path of the source instead of a SourceFile.  With this
change, a new SourceFile was built from the path.  However,
because the SourceFile's rendered html uses UUIDs for html ids
to ensure they are unique, the new SourceFile doesn't match an
existing one for the path and instead of the old one being removed
a duplicate was created.  Now the map of paths to SourceFiles is
used to delete the exact same SourceFile object in removeSource().