chore!: simplify language by removing some groups
The `Foundation:` and `Evaluates:` groups have been removed since
they are not needed and make the language more complicated.

A `Foundation:` is not needed since if a definition has a double
quote then it can be viewed as foundational.

The `Evaluates:` group is not needed since custom operators can
be created instead (such as `\real.+/` or `\integer.+/`).