chore!: remove the `where:` section
In all cases where the `where:` section is used, there is another
section that serves a similar purpose (for example,
`exists:where:suchThat:`, `forAll:where:suchThat:then:`,
`Theorem:given:where:if:then:`).

This causes confusion around what information should be put in the
`where:` section and what should be put in the other similar section
`suchThat:` or `if:`) making the language more complicated than it
needs to be.

As such, the `where:` section has been removed.  Instead, the other
related sections can be used.