client: EntitiesPanel is a higher-order component
This change is needed so that if multiple copies of an item is
shown in the panel (i.e. the same theorem is shown multiple times)
and one of them is closed, then only the item whose close button
was pressed should be removed from the panel.  Previously, every
copy of the item was removed from the panel.