feat: written sections can refer to defines target
For example, consider:
```
   [\function:on{A}:to{B}]
   Defines: f(x)
   when:
   .'A, B is \set'
   means:
   . forAll: a
     suchThat:
     .'a in A'
     then:
     .existsUnique: b
      suchThat:
     .'b in B'
     .'f(a) = b'
   written:
   . "f? : A?? \mapsto B??"
   called:
   . "function on A? to B?"
```
and notice the `f` is used in the `written:` section.  In this
case, `g is \function:on{X}:to{Y}` would be written as
`g : X \mapsto Y`.

Notice because `f` is used in the written section, then the `is`
is not used in the written form.  That is, it is `g : X \mapsto Y`
instead of `g is g : X \mapsto Y`.