feat: add support for placeholder variables
Placeholder variables are defined in square brackets or as
function parameters.  They differ from non-placeholder variables
in that they can be reused within a scope and shadow the previous
usage.  For example, in the following code:

```
Theorem:
given: f(x), g(x)
then: '\integral[x]{f(x) + g(x)} = ...'
```

The mappings `f` and `g` can use the placeholder variable (i.e.
we do not need to write `f(x)` and `g(y)`) and the introduction
of the placeholder variable in the square braces `[x]` can also
use `x` but shadows the usage of `x` in `f(x)` and `g(x)`.

It is an error for a placeholder and non-placeholder variable to
have the same name.