chore!: disable lhs rendering in `using:` section
Previously, a statement in a `using:` section of the form:
```
lhs := rhs
```
would have its left-hand-side rendered.  However, this breaks
displaying the entire statement if the left-hand-side contains
a backslash.

Now, the `lhs := rhs` is printed directly.  That is, it is
not rendered at all, so users can see what the alias is.