feat: `mlg` handles interrupted downloads
`mlg` first downloads the latest mathlingua jar to
`.bin/mathlingua.jar.part` and then moves that file to
`.bin/mathlingua.jar`.  This essentially ensures if the
download is cancelled before completing, the next run of `mlg`
will try the download again instead of trying to open a corrupt
jar file.