feat: add `have:by:` construct
The `have:by:` construct is used to specify that something is
true in a definition by having a reference to the theorem that
justifies it.

For example, `\rational.ordered.field` might have
```
...
means:
. '(Q, <) is \ordered.field'
...
```
where `Q` and `<` are defined previously.  However, that requires
justification.

Now, one can have
```
means:
. have: '(Q, <) is \ordered.field'
  by: '\some.theorem'
```
to point out what theorem justifies the definition.

Moreover, the `have:` section can have any clause.  That is, one
could have,
```
. have:
  . forAll: x, y
    suchThat: ...
    then: ...
  by: '\some.theorem'
```