client: fix side panel text appearance on ipad pro
On an iPad Pro, the text in the side panel would be shifted off
of the left side of the screen.  This has been corrected.

However, now the side panel text is aligned to the left of the
screen compared to being right next to the left side of the
page.