feat!: chalktalk sequence abstractions require {...}
Previously, a sequence-like abstraction in chalktalk could be
written as `{x_i}_i`.  Now they must be written as `{x_{i}}_{i}`.
Note the additional `{...}`.

This makes the syntax align with how sequence-like abstractions
are used in textalk and will enable names in chalktalk of the
form `x_1` which are supported names in textalk.