feat: `mlg` reports completions
An API endpoint and a `mlg` feature have been added to output
code completions that IDEs can use.

The API is used by the MathLingua frontend to provide completions,
while the `mlg` feature is used by the VS Code extension to get
completions that should be available in VS Code.

Specifically,
```
java -jar -D__MATHLINGUA_SHOW_COMPLETIONS__=true <mathlingua.jar> help
```
will cause a JSON string of the completions to be printed to
standard out in the form
```
interface Completions {
   items: Array<{ name: string; value: string }>;
}
```