feat!: `name?`, `name+?`, and `name-?` supported
Now if a `name?` appears in a `written:` or `called:` section,
the name will still be replaced with the appropriate match.

However, if a `name+?` appears, the name will be replaced with
a match enclosed in parens if the match is a complex expression.

Previously, the syntax `name??` was used and that syntax is no
longer supported.

Thus change was made to support the syntax `name-?` where a
match has any parentheses removed if it has them.

The use of `name-?` allows support for things such as `\frac{}{}`.
In particular, a signature that has a written as section with
a `\frac{}{}` can use `\frac{x-?}{y-?}` to ensure the numerator
and denominator do not have parentheses since they are not needed.