feat: add groundwork for Google Analytics support
The groundwork for being able to support Google Analytics
has been added, however it is not enabled.

Before doing so, support for configurations needs to be added.
With this, users can optionally specify a Google Analytics
number to be used by `mlg document`.  If present, Google Analytics
will be enabled using that number.  However, by default it is not
enabled.

Thus, the Google Analytics is only used by creators of a document
if and only if they decide they want to use it to understand
who is using their document.