feat: add support for Google Analytics
Currently, if the user sets the `GOOGLE_ANALYTICS_ID`
environment variable to a Google Analytics 4 Measurement ID, then
`mlg document` will use that id to enable Google Analytics in the
document it generates.

In that case, a banner will be shown on the bottom of the document
asking visitors if they agree to using cookies to allow the site
to analyze user traffic.  Only if the user clicks "I agree" will
the site record Google Analytics events.  Otherwise,
Google Analytics is not used at all.