client: render LaTeX twice
There appears to be a problem with LaTeX sometimes not being
rendered on initial page load.  In an attempt to address this,
now two calls to render the LaTeX are called on mounting a
component.

The performance impact should not be noticed by the user,
whereas not rendering the LaTeX severly impacts the usefulness
of the document being viewed.