chore: simplify the `mlg` runner
Now when `mlg update` is run, the `mlg` binary doesn't download
a new `mlg` binary and use a renaming trick to update the `mlg`
file in place.  This is done because doing so appears to make
Gatekeeper think the new `mlg` file is invalid on MacOS (even
though it is correctly notarized).

Instead, on update, the `mathlingua.jar` file is download and when
`mlg` is run, a hash of file is stored.  Then, if the user
downloads a new `mlg` binary directly, when it is run, the
`mathlingua.jar` file for the new `mlg` binary will be downloaded.