chore: remove `LocationTracker`
The `LocationTracker` usage was adding code complexitity and
wasn't actually needed.  It was only used so that row and column
information was not stored in the AST.  However, there really
isn't an issue doing that except in comparing two ASTs.  Thus,
the `LocationTracker` has been removed, greatly simplifying the
code.