feat: `Theorem:` has `where:suchThat:` sections
`Axiom:` and `Conjecture:` also have `where:suchThat:` sections.
This replaces the single `when:` section and aligns with
`forAll:where:suchThat:`, `exists:where:suchThat:`, and
`existsUnique:where:suchThat:`.

In particular, `is` and `in` statements appear in `where:`
and other statements appear in `suchThat:`.

In addition, a couple of issues were fixed:
* `Defines:satisfying:` commands can now be used in `is:` sections
* `generated:` sections render using the format used in the input
   source