rename formulation `parser.go` to `preparser.go`
The term `perparser` is used so that the `preparser.go` can have
a `PreParse` function that returns an expression with pseudo-nodes.

Then a follow-up pass can perform the shunting-yard algorithm to
consolidate those pseudo-nodes into single nodes.

This allows eash pass to be tested separately.  If, in the future,
having two passes is too slow, they can be consolidated at that
time after profiling has been done.